requirements.txt
discord.py
asycio